+++ src/debugadapter.ts
@@ -2478,18 +2478,18 @@ the value correspondents to a label.
 "-exec|e <<cmd>> <args>": cmd and args are directly passed to the remote (ZEsarUX, CSpect, ...). E.g. "-exec get-registers".
 "-help|h": This command. Do "-e help" to get all possible remote (ZEsarUX, CSpect, ...) commands.
 "-label|-l <XXX>": Returns the matching labels (XXX) with their values. Allows wildcard "*".
-"-md <address> <size> [dec|hex] [word] [little|big]": Memory dump at 'address' with 'size' bytes. Output is in 'hex' (default) or 'dec'imal. Per default data will be grouped in bytes.
+"-md [bank=<b>] <address> <size> [dec|hex] [word] [little|big]": Memory dump at 'address' with 'size' bytes. Output is in 'hex' (default) or 'dec'imal. Per default data will be grouped in bytes.
   But if chosen, words are output. Last argument is the endianness which is little endian by default.
-"-mdelta <address> <size> <string>": This does a 'delta' search on the address range. The memory is searched for byte
+"-mdelta [bank=<b>] <address> <size> <string>": This does a 'delta' search on the address range. The memory is searched for byte
   sequences that contain the same deltas as the given string. E.g. if you would search for "MIK" not only "MIK"
   would be found but also "mik" or "njl". The whole memory is dumped but all values are adjusted by an offset to
   match the searched string. If several sequences are found the memory might be dumped several times.
   The idea is to find strings also if they are not coded as ASCII sequence but with some other, unknown coding
   scheme.
 "-memmodel": Prints slot and bank info of the currently used memory model.
-"-ml <address> <filepath>": Loads a binary file into memory. The filepath is relative to the TMP directory.
-"-ms <address> <size> <filename>": Saves a memory dump to a file. The file is saved to the temp directory.
-"-msetb <address> <value> [repeat]":
+"-ml [bank=<b>] <address> <filepath>": Loads a binary file into memory. The filepath is relative to the TMP directory.
+"-ms [bank=<b>] <address> <size> <filename>": Saves a memory dump to a file. The file is saved to the temp directory.
+"-msetb [bank=<b>] <address> <value> [repeat]":
 	- address: The address to fill. Can also be a label or expression.
 	- value: The byte value to set.
 	- repeat: (Optional) How often the value is repeated.
@@ -2497,7 +2497,7 @@ the value correspondents to a label.
 	"-msetb <8000h> <0Fh>" : Puts a 15 into memory location 0x8000.
 	"-msetb <8000h> <0> <100h>" : fills memory locations 0x8000 to 0x80FF with zeroes.
 	"-msetb <fill_colors_ptr+4> <FEh>": If fill_colors_ptr is e.g. 0xCF02 the value FEh is put into location 0xCF06.
-"-msetw <address> <value> [repeat [endianness]]:"
+"-msetw [bank=<b>] <address> <value> [repeat [endianness]]:"
 	- address: The address to fill. Can also be a label or expression.
 	- value: The word value to set.
 	- repeat: (Optional) How often the value is repeated.